+++ src/students/example/mlflow/era5_temperature_project/era5_pipeline/resources.py
@@ -3,6 +3,7 @@ import os
 from mlflow.tracking import MlflowClient
 from cdsapi import Client
 import dagster as dg
+import pydantic as pyd
 
 # Configuration for Local SQLite and Local Artifacts
 # Using DAGSTER_HOME if set, otherwise, defaults to the current directory
@@ -38,3 +39,39 @@ class CDSAPI(dg.ConfigurableResource):
     @property
     def client(self):
         return Client(url=self.host_url, key=self.api_key)
+
+
+# configuration for the raw_netcdf_dataset asset
+class Era5RequestConfig(dg.Config):
+    product_type: str = pyd.Field(
+        default="reanalysis",
+        description="The product type to request"
+    )
+    variable: str = pyd.Field(
+        default="2m_temperature",
+        description="The meteorological variable to retrieve"
+    )
+    year: str = pyd.Field(
+        default="2023",
+        description="The year for which to retrieve data"
+    )
+    month: str = pyd.Field(
+        default="01",
+        description="The month for which to retrieve data"
+    )
+    day: list[str] = pyd.Field(
+        default=[f"{i:02d}" for i in range(1, 16)],
+        description="A list of days to retrieve"
+    )
+    time: list[str] = pyd.Field(
+        default=["00:00", "06:00", "12:00", "18:00"],
+        description="Times of day to retrieve data"
+    )
+    area: list[float] = pyd.Field(
+        default=[50.0, -5.0, 45.0, 5.0],
+        description="Area: [North, West, South, East]"
+    )
+    format: str = pyd.Field(
+        default="netcdf",
+        description="Format to download (e.g., netcdf)"
+    )
